+++ x/btcstaking/keeper/finality_providers.go
@@ -56,7 +56,7 @@ func (k Keeper) AddFinalityProvider(goCtx context.Context, msg *types.MsgCreateF
 		}
 		// Ensure there's an IBC channel open if it is a Cosmos BSN
 		if cr.GetCosmosConsumerMetadata() != nil {
-			hasChannel := k.BscKeeper.ConsumerHasIBCChannelOpen(ctx, bsnID)
+			hasChannel := k.BscKeeper.ConsumerHasIBCChannelOpen(ctx, bsnID, cr.GetCosmosConsumerMetadata().ChannelId)
 			if !hasChannel {
 				return types.ErrFpConsumerNoIBCChannelOpen
 			}